refactor out FieldTypeHandler & add type support
